// archive/sdk-pre-vnext-20260606/src/runtime/generated/runtime/v1/auth.ts
export interface AppModeManifest {
    /**
     * @generated from protobuf field: nimi.runtime.v1.AppMode app_mode = 1
     */
    appMode: AppMode;
    /**
     * @generated from protobuf field: bool runtime_required = 2
     */
    runtimeRequired: boolean;
    /**
     * @generated from protobuf field: bool realm_required = 3
     */
    realmRequired: boolean;
    /**
     * @generated from protobuf field: nimi.runtime.v1.WorldRelation world_relation = 4
     */
    worldRelation: WorldRelation;
}
/**
 * @generated from protobuf message nimi.runtime.v1.RegisterAppRequest
 */
export interface RegisterAppRequest {
    /**
     * @generated from protobuf field: string app_id = 1
     */
    appId: string;
    /**
     * @generated from protobuf field: string app_instance_id = 2
     */
    appInstanceId: string;
    /**
     * @generated from protobuf field: string device_id = 3
     */
    deviceId: string;
    /**
     * @generated from protobuf field: string app_version = 4
     */
    appVersion: string;
    /**
     * @generated from protobuf field: repeated string capabilities = 5
     */
    capabilities: string[];
    /**
     * @generated from protobuf field: nimi.runtime.v1.AppModeManifest mode_manifest = 6
     */
    modeManifest?: AppModeManifest;
    /**
     * K-AUTHSVC-014: explicit local developer-registration intent. When true AND
     * the daemon's auth.developerRegistration.enabled gate is on, RegisterApp may
     * admit an app_id not present in the Nimi App registry. Default false keeps
     * production admission fail-closed (APP_NOT_REGISTERED).
     *
     * @generated from protobuf field: bool developer_registration = 7
     */
    developerRegistration: boolean;
}
/**
 * @generated from protobuf message nimi.runtime.v1.RegisterAppResponse
 */
export interface RegisterAppResponse {
    /**
     * @generated from protobuf field: string app_instance_id = 1
     */
    appInstanceId: string;
    /**
     * @generated from protobuf field: bool accepted = 2
     */
    accepted: boolean;
    /**
     * @generated from protobuf field: nimi.runtime.v1.ReasonCode reason_code = 3
     */
    reasonCode: ReasonCode;
}
// ...
export enum AppMode {
    /**
     * @generated from protobuf enum value: APP_MODE_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: APP_MODE_LITE = 1;
     */
    LITE = 1,
    /**
     * @generated from protobuf enum value: APP_MODE_CORE_ONLY = 2;
     */
    CORE_ONLY = 2,
    /**
     * @generated from protobuf enum value: APP_MODE_FULL = 3;
     */
    FULL = 3
}
/**
 * @generated from protobuf enum nimi.runtime.v1.WorldRelation
 */
export enum WorldRelation {
    /**
     * @generated from protobuf enum value: WORLD_RELATION_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: WORLD_RELATION_NONE = 1;
     */
    NONE = 1,
    /**
     * @generated from protobuf enum value: WORLD_RELATION_RENDER = 2;
     */
    RENDER = 2,
    /**
     * @generated from protobuf enum value: WORLD_RELATION_EXTENSION = 3;
     */
    EXTENSION = 3
}
// ...
export const AppModeManifest = new AppModeManifest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class RegisterAppRequest$Type extends MessageType<RegisterAppRequest> {
    constructor() {
        super("nimi.runtime.v1.RegisterAppRequest", [
            { no: 1, name: "app_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "app_instance_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "device_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "app_version", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 5, name: "capabilities", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ },
            { no: 6, name: "mode_manifest", kind: "message", T: () => AppModeManifest },
            { no: 7, name: "developer_registration", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
    create(value?: PartialMessage<RegisterAppRequest>): RegisterAppRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.appId = "";
        message.appInstanceId = "";
        message.deviceId = "";
        message.appVersion = "";
        message.capabilities = [];
        message.developerRegistration = false;
        if (value !== undefined)
            reflectionMergePartial<RegisterAppRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: RegisterAppRequest): RegisterAppRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string app_id */ 1:
                    message.appId = reader.string();
                    break;
                case /* string app_instance_id */ 2:
                    message.appInstanceId = reader.string();
                    break;
                case /* string device_id */ 3:
                    message.deviceId = reader.string();
                    break;
                case /* string app_version */ 4:
                    message.appVersion = reader.string();
                    break;
                case /* repeated string capabilities */ 5:
                    message.capabilities.push(reader.string());
                    break;
                case /* nimi.runtime.v1.AppModeManifest mode_manifest */ 6:
                    message.modeManifest = AppModeManifest.internalBinaryRead(reader, reader.uint32(), options, message.modeManifest);
                    break;
                case /* bool developer_registration */ 7:
                    message.developerRegistration = reader.bool();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: RegisterAppRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string app_id = 1; */
        if (message.appId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.appId);
        /* string app_instance_id = 2; */
        if (message.appInstanceId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.appInstanceId);
        /* string device_id = 3; */
        if (message.deviceId !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.deviceId);
        /* string app_version = 4; */
        if (message.appVersion !== "")
            writer.tag(4, WireType.LengthDelimited).string(message.appVersion);
        /* repeated string capabilities = 5; */
        for (let i = 0; i < message.capabilities.length; i++)
            writer.tag(5, WireType.LengthDelimited).string(message.capabilities[i]);
        /* nimi.runtime.v1.AppModeManifest mode_manifest = 6; */
        if (message.modeManifest)
            AppModeManifest.internalBinaryWrite(message.modeManifest, writer.tag(6, WireType.LengthDelimited).fork(), options).join();
        /* bool developer_registration = 7; */
        if (message.developerRegistration !== false)
            writer.tag(7, WireType.Varint).bool(message.developerRegistration);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message nimi.runtime.v1.RegisterAppRequest
 */
export const RegisterAppRequest = new RegisterAppRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class RegisterAppResponse$Type extends MessageType<RegisterAppResponse> {
    constructor() {
        super("nimi.runtime.v1.RegisterAppResponse", [
            { no: 1, name: "app_instance_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "accepted", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 3, name: "reason_code", kind: "enum", T: () => ["nimi.runtime.v1.ReasonCode", ReasonCode] }
        ]);
    }
    create(value?: PartialMessage<RegisterAppResponse>): RegisterAppResponse {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.appInstanceId = "";
        message.accepted = false;
        message.reasonCode = 0;
        if (value !== undefined)
            reflectionMergePartial<RegisterAppResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: RegisterAppResponse): RegisterAppResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string app_instance_id */ 1:
                    message.appInstanceId = reader.string();
                    break;
                case /* bool accepted */ 2:
                    message.accepted = reader.bool();
                    break;
                case /* nimi.runtime.v1.ReasonCode reason_code */ 3:
                    message.reasonCode = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: RegisterAppResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string app_instance_id = 1; */
        if (message.appInstanceId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.appInstanceId);
        /* bool accepted = 2; */
        if (message.accepted !== false)
            writer.tag(2, WireType.Varint).bool(message.accepted);
        /* nimi.runtime.v1.ReasonCode reason_code = 3; */
        if (message.reasonCode !== 0)
            writer.tag(3, WireType.Varint).int32(message.reasonCode);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message nimi.runtime.v1.RegisterAppResponse
 */
export const RegisterAppResponse = new RegisterAppResponse$Type();